refactor: update reducer, move selectors to /user/ and /data/ modules
