--- src/reducer/cities/cities.js
@@ -1,46 +1,0 @@
-const initialState = {
-  city: `Amsterdam`,
-};
-
-
-const ActionType = {
-  CHANGE_CITY: `CHANGE_CITY`,
-};
-
-
-/**
- * Methods that returns actions for reducer (Object with type and payload params)
- * @return {Object}
- */
-const ActionCreator = {
-  changeCity: (city) => ({
-    type: ActionType.CHANGE_CITY,
-    payload: city
-  }),
-};
-
-
-/**
- * Reducer for change application state
- * @param {Object} state [state = initialState]
- * @param {Object} action
- * @param {String} action.type
- * @return {Object}
- */
-const reducer = (state = initialState, action) => {
-  switch (action.type) {
-    case ActionType.CHANGE_CITY:
-      return Object.assign({}, state, {
-        city: action.payload
-      });
-  }
-
-  return state;
-};
-
-
-export {
-  ActionCreator,
-  ActionType,
-  reducer
-};
