+++ src/reducer/name-spaces.js
@@ -1,5 +1,4 @@
 const NameSpace = {
-  CITIES: `CITIES`,
   DATA: `DATA`,
   USER: `USER`
 };
